add/error_cases
