+++ map/error/design_missing_20_13.cub
@@ -1,7 +1,6 @@
 NO ./texture/manf1.xpm
 SO ./texture/manb1.xpm
 WE ./texture/ove.xpm 
-EA ./texture/castle.xpm
 
 F 118,118,118
 C 90,90,90
@@ -12,9 +11,9 @@ C 90,90,90
 10001010101010101001
 10000000000000000001
 10000000000000001001
+1000000000N000001001
 10000000000000001001
-1000000000000001001S
-10000000000001111001111
+10000000000001111001
 10000000000000000001
 10000000000000000001
 10000000000000000001
