+++ map/error/open_unable_20_13.cub
@@ -12,9 +12,9 @@ C 90,90,90
 10001010101010101001
 10000000000000000001
 10000000000000001001
+1000000000N000001001
 10000000000000001001
-1000000000000001001S
-10000000000001111001111
+10000000000001111001
 10000000000000000001
 10000000000000000001
 10000000000000000001
