+++ map/error/rgb_decimal_20_13.cub
